Move DocumentHead element to invite-accept
